requirements.txt
annotated-types==0.7.0
azure-functions==1.24.0
MarkupSafe==3.0.3
pydantic==2.12.4
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
